MAINT: seperate tests for readability
